requirements.txt
Django
djangorestframework
stripe
python-dotenv
psycopg2-binary
gunicorn
